update: балансы теперь хранят в копейках, убраны Nonnull'ы из-за их бесполезности и другие исправления

+++ src/main/java/ru/stepagin/becoder/component/SetupDataLoader.java
@@ -39,7 +39,7 @@ public class SetupDataLoader implements ApplicationListener<ContextRefreshedEven
         LegalAccountDTO dto = accountService.createAccount(admin);
         LegalAccountEntity account = new LegalAccountEntity();
         account.setId(UUID.fromString(dto.getId()));
-        account.setBalance(dto.getBalance());
+        account.setBalance((long) dto.getBalance());
 
 
     }
